server/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Cors==4.0.0
pytest==7.4.2
gunicorn==21.2.0
